cmdpart正式下岗
