--- GTMBplugin_B/Parts/customcmds/customcmds.part
@@ -1,11 +1,0 @@
-{
-    "bpType": "customcmdsPart",
-    "changes": {},
-    "classType": "PartInstance",
-    "clientPartType": null,
-    "customVariables": [],
-    "directoryPath": "customcmds",
-    "metaType": "customcmdsPartMeta",
-    "objId": "ce01f37f3e394aa2a2d7a871a0ddb226",
-    "partType": "customcmdsPart"
-}
